refactor: replace Double with BigDecimal for all monetary values
- User.balance and Transaction.amount now use BigDecimal with DECIMAL(15,2)
- All DTOs updated (RegisterRequest, DepositRequest, WithdrawRequest, TransferRequest)
- UserService arithmetic migrated to BigDecimal methods (add, subtract, compareTo)
- calculateFee() returns BigDecimal
- Tests updated to use BigDecimal literals
- DB migration already uses DECIMAL(15,2) - no schema change needed

+++ src/main/java/com/mpesa/dto/TransferRequest.java
@@ -1,6 +1,7 @@
 package com.mpesa.dto;
 
 import jakarta.validation.constraints.*;
+import java.math.BigDecimal;
 
 public class TransferRequest {
     @NotNull(message = "Sender ID is required")
@@ -13,7 +14,7 @@ public class TransferRequest {
     @NotNull(message = "Amount is required")
     @DecimalMin(value = "0.01", message = "Amount must be positive")
     @DecimalMax(value = "500000", message = "Amount exceeds maximum limit")
-    private Double amount;
+    private BigDecimal amount;
 
     @NotBlank(message = "PIN is required")
     private String pin;
@@ -22,8 +23,8 @@ public class TransferRequest {
     public void setFromUserId(Integer fromUserId) { this.fromUserId = fromUserId; }
     public Integer getToUserId() { return toUserId; }
     public void setToUserId(Integer toUserId) { this.toUserId = toUserId; }
-    public Double getAmount() { return amount; }
-    public void setAmount(Double amount) { this.amount = amount; }
+    public BigDecimal getAmount() { return amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
     public String getPin() { return pin; }
     public void setPin(String pin) { this.pin = pin; }
 }
